+++ source/renderer/zettlr-quicklook.js
@@ -19,7 +19,7 @@ const path = require('path');
 // First codemirror addons
 require('codemirror/addon/mode/overlay');
 require('codemirror/addon/edit/continuelist');
-require('codemirror/addon/edit/indentlist');
+require('./assets/codemirror/indentlist.js');
 require('codemirror/addon/search/search');
 require('codemirror/addon/search/searchcursor');
 require('codemirror/addon/search/jump-to-line');